+++ app/templates/sources/src/main/groovy/package/ProjectExtension.groovy
@@ -1,11 +1,14 @@
 package <%= projectPackage %>
 
+import groovy.transform.CompileStatic
+
 /**
  * <%= pluginName %> plugin extension.
  *
  * @author <%= authorName %>
  * @since <%= date %>
  */
+@CompileStatic
 class <%= pluginClassPrefix %>Extension {
 
     String foo