+++ app/templates/sources/src/main/groovy/package/ProjectPlugin.groovy
@@ -1,5 +1,7 @@
 package <%= projectPackage %>
 
+import groovy.transform.CompileStatic
+import groovy.transform.TypeCheckingMode
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.GradleException
@@ -10,9 +12,11 @@ import org.gradle.api.GradleException
  * @author <%= authorName %>
  * @since <%= date %>
  */
+@CompileStatic
 class <%= pluginClassPrefix %>Plugin implements Plugin<Project> {
 
     @Override
+    @CompileStatic(TypeCheckingMode.SKIP)
     void apply(Project project) {
         <%= pluginClassPrefix %>Extension extension = project.extensions.create('<%= pluginExtensionName %>', <%= pluginClassPrefix %>Extension)